// manager/src/store/actions/types.ts
export enum AuthActionType {
  CHANGE_PASSWORD = "CHANGE_PASSWORD",
  CHANGE_EMAIL = "CHANGE_EMAIL",
  LOGIN_USER = "LOGIN_USER",
  LOGIN_FAIL = "LOGIN_FAIL",
  SET_LOADING = "SET_LOADING",
}

export enum EmployeeFormActionType {
  UPDATE_FORM = "UPDATE_FORM",
  CLEAR_FORM = "CLEAR_FORM",
  INIT_FORM = "INIT_FORM",
}

export enum EmployeeActionType {
  SET_EMPLOYEES = "SET_EMPLOYEES"
}